+++ parasol_cpu/tests/e2e_tests/vector_add.rs
@@ -17,7 +17,8 @@ fn can_run_from_elf() {
 
     let mut proc = FheComputer::new(&enc, &eval);
 
-    let data = std::array::from_fn::<_, 8, _>(|i| UInt::<8, _>::encrypt_secret(i as u64, &enc, sk));
+    let data =
+        std::array::from_fn::<_, 8, _>(|i| UInt::<8, _>::encrypt_secret(i as u128, &enc, sk));
 
     let a = memory.try_allocate_type(&data).unwrap();
     let b = memory.try_allocate_type(&data).unwrap();
